requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
fastapi
uvicorn[standard]
python-multipart
pillow
numpy
opencv-python-headless
timm
albumentations
torch
torchvision
